added changes to navbar
