+++ src/App.js
@@ -5,7 +5,7 @@ import {Team} from "./pages/Team"
 import {BrowserRouter,Routes,Route} from "react-router-dom"
 import {Achievements} from "./pages/Achievements"
 import {Projects} from "./pages/Projects"
-import Navbar from "./components/Navbar"
+import Navbar from "./components/navbar/Navbar"
 
 
 const App = () => {
