+++ src/components/navbar/data.js
@@ -1,0 +1,21 @@
+const linksData = [
+    {
+      name: 'HOME',
+      url: '/',
+    },
+    {
+      name: 'TEAMS',
+      url: '/teams',
+    },
+    {
+      name: 'PROJECTS',
+      url: '/projects',
+    },
+    {
+      name: 'ACHIEVEMENTS',
+      url: '/achievements',
+    },
+  ];
+  
+  export default linksData;
+
