+++ src/index.css
@@ -15,7 +15,7 @@ body {
 
 
 .glass{
-  background: linear-gradient(135deg, rgba(255, 255, 255, 0.2), rgba(255, 255, 255, 0));
+  background: linear-gradient(135deg, rgba(255, 255, 255, 0.2), rgba(255, 255, 255, 0.1));
     backdrop-filter: blur(10px);
     -webkit-backdrop-filter: blur(10px);
     border-radius: 20px;
@@ -24,29 +24,6 @@ body {
     z-index: 100px;
   
 }
-
-
-
-
-
-
-.link.active {
-  background-color: #ff6600; 
-  font-weight: bold;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 .hoverable{
@@ -67,15 +44,19 @@ body {
   content: '';
   top: 100%;
   left: 5%;
-  height:10px;
+  /* height:5px; */
   width:90%;
   opacity:0;
   background: -webkit-radial-gradient(center, ellipse, rgba(255, 255, 255, 0.35) 0%, rgba(255, 255, 255, 0) 80%);
-  background: radial-gradient(ellipse at center, rgba(255, 255, 255, 0.35) 0%, rgba(255, 255, 255, 0) 80%);
+  background: radial-gradient(ellipse at center, rgba(255, 255, 255, 0.2) 0%, rgba(255, 255, 255, 0) 80%);
   /* W3C */
   transition-duration: 0.3s;
   transition-property: transform, opacity;
+  
 }
+
+
+
 
 
 
@@ -114,3 +95,18 @@ body {
 }
 
 
+
+.neon{
+  color:#fff;
+  text-shadow:0 0 20px #fd700a;
+
+
+}
+
+.neon:before{
+  
+  background:#fd700a;
+  z-index:-2;
+  opacity:.4;
+  filter:blur(800px);
+}
